UTC calculation error fixed
Error in UTC calculation fixed. Error occured when UTC changes to next day/month/year and your local-time is e.g. two hours "behind" , so you are still in the previous day/month/year. You got a negative hour in the result and the plugin crashed or the date was not updated.
Also added a long-output for date and time.